+++ package.json
@@ -17,26 +17,26 @@
     "seed": "tsx prisma/seed.ts"
   },
   "dependencies": {
-    "@aws-sdk/client-sesv2": "3.835.0",
-    "@hookform/resolvers": "5.1.1",
-    "@prisma/client": "6.10.1",
+    "@aws-sdk/client-sesv2": "3.857.0",
+    "@hookform/resolvers": "5.2.1",
+    "@prisma/client": "6.13.0",
     "bcryptjs": "3.0.2",
     "next": "15.3.4",
     "next-auth": "5.0.0-beta.28",
-    "react": "19.1.0",
-    "react-dom": "19.1.0",
-    "react-hook-form": "7.58.1",
-    "zod": "3.25.67"
+    "react": "19.1.1",
+    "react-dom": "19.1.1",
+    "react-hook-form": "7.61.1",
+    "zod": "4.0.14"
   },
   "devDependencies": {
-    "@biomejs/biome": "2.0.4",
-    "@tailwindcss/postcss": "4.1.10",
-    "@types/node": "20.19.1",
-    "@types/react": "19.1.8",
-    "@types/react-dom": "19.1.6",
-    "daisyui": "5.0.43",
-    "prisma": "6.10.1",
-    "tailwindcss": "4.1.10",
+    "@biomejs/biome": "2.1.3",
+    "@tailwindcss/postcss": "4.1.11",
+    "@types/node": "22.17.0",
+    "@types/react": "19.1.9",
+    "@types/react-dom": "19.1.7",
+    "daisyui": "5.0.50",
+    "prisma": "6.13.0",
+    "tailwindcss": "4.1.11",
     "tsx": "4.20.3",
     "typescript": "5.8.3"
   },
