+++ src/lib/zod.ts
@@ -1,4 +1,4 @@
-import * as z from "zod/v4";
+import * as z from "zod";
 
 const passwordSchema = z.string().min(8).max(32);
